requirements.txt
openai
streamlit
python-dotenv
pdfplumber
python-docx
